+++ payment_error.php
@@ -14,7 +14,6 @@ include(dirname(__FILE__) . '/../../config/config.inc.php');
 include(dirname(__FILE__) . '/../../header.php');
 
 $smarty->assign(array('this_path' => __PS_BASE_URI__));
-$smarty->display(_PS_MODULE_DIR_ . 'servired/views/templates/frontend/payment_error.tpl');
-
+$smarty->display(_PS_MODULE_DIR_ . 'servired/views/templates/hook/payment_error.tpl');
 include(dirname(__FILE__) . '/../../footer.php');
 ?>